fix window minimize bug fix pop-up window conflicts with main window
